implemented keycloak, docker, and feign client in project

+++ category-service/src/main/java/com/category/controller/SalonCategoryController.java
@@ -3,6 +3,8 @@ package com.category.controller;
 import com.category.model.Category;
 import com.category.payload.DTO.SalonDTO;
 import com.category.service.CategoryService;
+import com.category.service.client.SalonFeignClient;
+import org.apache.tomcat.util.http.parser.Authorization;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -15,21 +17,27 @@ public class SalonCategoryController {
     @Autowired
     CategoryService categoryService;
 
-    @PostMapping("/create")
-    public ResponseEntity<Category> saveCategory(@RequestBody Category category){
-       SalonDTO salonDTO = new SalonDTO();
+    @Autowired
+    SalonFeignClient salonFeignClient;
 
-        salonDTO.setId(1L);
+    @PostMapping("/create")
+    public ResponseEntity<Category> saveCategory(@RequestBody Category category,
+                                                 @RequestHeader("Authorization") String jwt) throws Exception {
+       SalonDTO salonDTO = salonFeignClient.getSalonByOwnerId(jwt).getBody();
         Category category1 = categoryService.saveCategory(category, salonDTO);
 
         return new ResponseEntity<>(category1, HttpStatus.CREATED);
     }
 
     @DeleteMapping("/delete")
-    public ResponseEntity<String> deleteCategory(@PathVariable Long id) throws Exception {
+    public ResponseEntity<String> deleteCategory(@PathVariable Long id,
+                                                 @RequestHeader("Authorization") String jwt) throws Exception {
 
-        SalonDTO salonDTO = new SalonDTO();
-        salonDTO.setId(1L);
+        SalonDTO salonDTO = salonFeignClient.getSalonByOwnerId(jwt).getBody();
+
+        if(salonDTO == null){
+            throw new Exception("user not found "+ salonDTO);
+        }
 
         categoryService.deleteCategoryById(id, salonDTO.getId());
